+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/container_service_client.py
@@ -78,7 +78,7 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
         _PROFILE_TAG: {
             'open_shift_managed_clusters': '2018-09-30-preview',
             'container_services': '2017-07-01',
-            'managed_clusters': '2019-10-01',
+            'managed_clusters': '2020-01-01',
             'operations': '2018-03-31',
             None: DEFAULT_API_VERSION
         }},
@@ -112,6 +112,7 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
            * 2019-04-01: :mod:`v2019_04_01_.models<azure.mgmt.containerservice.v2019_04_01.models>`
            * 2019-08-01: :mod:`v2019_08_01_.models<azure.mgmt.containerservice.v2019_08_01.models>`
            * 2019-10-01: :mod:`v2019_10_01_.models<azure.mgmt.containerservice.v2019_10_01.models>`
+           * 2020-01-01: :mod:`v2020_01_01_.models<azure.mgmt.containerservice.v2020_01_01.models>`
         """
         if api_version == '2017-07-01':
             from .v2017_07_01 import models
@@ -137,6 +138,9 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
         elif api_version == '2019-10-01':
             from .v2019_10_01 import models
             return models
+        elif api_version == '2020-01-01':
+            from .v2020_01_01 import models
+            return models
         raise NotImplementedError("APIVersion {} is not available".format(api_version))
 
     @property
@@ -147,6 +151,7 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
            * 2019-04-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2019_04_01.operations.AgentPoolsOperations>`
            * 2019-08-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2019_08_01.operations.AgentPoolsOperations>`
            * 2019-10-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2019_10_01.operations.AgentPoolsOperations>`
+           * 2020-01-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2020_01_01.operations.AgentPoolsOperations>`
         """
         api_version = self._get_api_version('agent_pools')
         if api_version == '2019-02-01':
@@ -157,6 +162,8 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
             from .v2019_08_01.operations import AgentPoolsOperations as OperationClass
         elif api_version == '2019-10-01':
             from .v2019_10_01.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2020-01-01':
+            from .v2020_01_01.operations import AgentPoolsOperations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
@@ -178,6 +185,9 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
         elif api_version == '2019-10-01':
             from .v2019_04_01.operations import ContainerServicesOperations as OperationClass
             api_version = "2019-04-01" #2019-10-01 generated Python SDK contains no list_orchestrator op
+        elif api_version == '2020-01-01':
+            from .v2019_04_01.operations import ContainerServicesOperations as OperationClass
+            api_version = "2019-04-01" #2020-01-01 generated Python SDK contains no list_orchestrator op
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
@@ -192,6 +202,7 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
            * 2019-04-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2019_04_01.operations.ManagedClustersOperations>`
            * 2019-08-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2019_08_01.operations.ManagedClustersOperations>`
            * 2019-10-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2019_10_01.operations.ManagedClustersOperations>`
+           * 2020-01-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2020_01_01.operations.ManagedClustersOperations>`
         """
         api_version = self._get_api_version('managed_clusters')
         if api_version == '2018-03-31':
@@ -206,6 +217,8 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
             from .v2019_08_01.operations import ManagedClustersOperations as OperationClass
         elif api_version == '2019-10-01':
             from .v2019_10_01.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2020-01-01':
+            from .v2020_01_01.operations import ManagedClustersOperations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
@@ -233,6 +246,7 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
            * 2019-04-01: :class:`Operations<azure.mgmt.containerservice.v2019_04_01.operations.Operations>`
            * 2019-08-01: :class:`Operations<azure.mgmt.containerservice.v2019_08_01.operations.Operations>`
            * 2019-10-01: :class:`Operations<azure.mgmt.containerservice.v2019_10_01.operations.Operations>`
+           * 2020-01-01: :class:`Operations<azure.mgmt.containerservice.v2020_01_01.operations.Operations>`
         """
         api_version = self._get_api_version('operations')
         if api_version == '2018-03-31':
@@ -247,6 +261,8 @@ class ContainerServiceClient(MultiApiClientMixin, SDKClient):
             from .v2019_08_01.operations import Operations as OperationClass
         elif api_version == '2019-10-01':
             from .v2019_10_01.operations import Operations as OperationClass
+        elif api_version == '2020-01-01':
+            from .v2020_01_01.operations import Operations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))