+++ src/azure-firewall/azext_firewall/commands.py
@@ -13,7 +13,7 @@ from ._client_factory import cf_firewalls, cf_firewall_fqdn_tags, cf_firewall_po
 from ._util import (
     list_network_resource_property, get_network_resource_property_entry, delete_network_resource_property_entry)
 
-from ._validators import validate_af_network_rule, validate_af_nat_rule
+from ._validators import validate_af_network_rule, validate_af_nat_rule, validate_af_application_rule
 
 
 # pylint: disable=too-many-locals, too-many-statements
@@ -58,6 +58,12 @@ def load_command_table(self, _):
         g.show_command('show')
         g.generic_update_command('update', custom_func_name='update_azure_firewall')
 
+    with self.command_group('network firewall threat-intel-whitelist', network_firewall_sdk, is_preview=True, min_api='2019-09-01') as g:
+        g.custom_command('create', 'create_azure_firewall_threat_intel_whitelist')
+        g.custom_command('delete', 'delete_azure_firewall_threat_intel_whitelist')
+        g.custom_command('show', 'show_azure_firewall_threat_intel_whitelist')
+        g.generic_update_command('update', custom_func_name='update_azure_firewall_threat_intel_whitelist')
+
     with self.command_group('network firewall ip-config', network_util) as g:
         g.custom_command('create', 'create_af_ip_configuration')
         g.command('list', list_network_resource_property('azure_firewalls', 'ip_configurations'))
@@ -67,7 +73,7 @@ def load_command_table(self, _):
     af_rules = {
         'network_rule': {'scope': 'network-rule', 'validator': validate_af_network_rule},
         'nat_rule': {'scope': 'nat-rule', 'validator': validate_af_nat_rule},
-        'application_rule': {'scope': 'application-rule', 'validator': None}
+        'application_rule': {'scope': 'application-rule', 'validator': validate_af_application_rule}
     }
 
     for rule_type, af_rule in af_rules.items():