+++ src/eventgrid/azext_eventgrid/vendored_sdks/eventgrid/models/operation_paged.py
@@ -14,7 +14,7 @@ from msrest.paging import Paged
 
 class OperationPaged(Paged):
     """
-    A paging container for iterating over a list of :class:`Operation <microsoft.azure.management.eventgrid.models.Operation>` object
+    A paging container for iterating over a list of :class:`Operation <azext_eventgrid.models.Operation>` object
     """
 
     _attribute_map = {