+++ src/image-copy/azext_imagecopy/create_target.py
@@ -9,6 +9,9 @@ import time
 from knack.util import CLIError
 from knack.log import get_logger
 
+from msrestazure.tools import resource_id
+from azure.cli.core.commands.client_factory import get_subscription_id
+
 from azext_imagecopy.cli_utils import run_cli_command, prepare_cli_command
 
 logger = get_logger(__name__)
@@ -18,7 +21,7 @@ STORAGE_ACCOUNT_NAME_LENGTH = 24
 
 # pylint: disable=too-many-statements
 # pylint: disable=too-many-locals
-def create_target_image(location, transient_resource_group_name, source_type, source_object_name,
+def create_target_image(cmd, location, transient_resource_group_name, source_type, source_object_name,
                         source_os_disk_snapshot_name, source_os_disk_snapshot_url, source_os_type,
                         target_resource_group_name, azure_pool_frequency, tags, target_name, target_subscription,
                         export_as_snapshot, timeout):
@@ -27,7 +30,7 @@ def create_target_image(location, transient_resource_group_name, source_type, so
         STORAGE_ACCOUNT_NAME_LENGTH - len(location))
 
     # create the target storage account. storage account name must be lowercase.
-    logger.warn(
+    logger.warning(
         "%s - Creating target storage account (can be slow sometimes)", location)
     target_storage_account_name = location.lower() + random_string
     cli_cmd = prepare_cli_command(['storage', 'account', 'create',
@@ -70,7 +73,7 @@ def create_target_image(location, transient_resource_group_name, source_type, so
     logger.debug("sas token: %s", sas_token)
 
     # create a container in the target blob storage account
-    logger.warn(
+    logger.warning(
         "%s - Creating container in the target storage account", location)
     target_container_name = 'snapshots'
     cli_cmd = prepare_cli_command(['storage', 'container', 'create',
@@ -82,7 +85,7 @@ def create_target_image(location, transient_resource_group_name, source_type, so
 
     # Copy the snapshot to the target region using the SAS URL
     blob_name = source_os_disk_snapshot_name + '.vhd'
-    logger.warn(
+    logger.warning(
         "%s - Copying blob to target storage account", location)
     cli_cmd = prepare_cli_command(['storage', 'blob', 'copy', 'start',
                                    '--source-uri', source_os_disk_snapshot_url,
@@ -100,10 +103,10 @@ def create_target_image(location, transient_resource_group_name, source_type, so
                                  azure_pool_frequency, location, target_subscription)
     msg = "{0} - Copy time: {1}".format(
         location, datetime.datetime.now() - start_datetime)
-    logger.warn(msg)
+    logger.warning(msg)
 
     # Create the snapshot in the target region from the copied blob
-    logger.warn(
+    logger.warning(
         "%s - Creating snapshot in target region from the copied blob", location)
     target_blob_path = target_blob_endpoint + \
         target_container_name + '/' + blob_name
@@ -113,11 +116,21 @@ def create_target_image(location, transient_resource_group_name, source_type, so
     else:
         snapshot_resource_group_name = transient_resource_group_name
 
+    storage_account_name = target_blob_path.split('.')[0].split('/')[-1]
+    if target_subscription:
+        subscription_id = target_subscription
+    else:
+        subscription_id = get_subscription_id(cmd.cli_ctx)
+    source_storage_account_id = resource_id(
+        subscription=subscription_id, resource_group=transient_resource_group_name,
+        namespace='Microsoft.Storage', type='storageAccounts', name=storage_account_name)
+
     cli_cmd = prepare_cli_command(['snapshot', 'create',
                                    '--resource-group', snapshot_resource_group_name,
                                    '--name', target_snapshot_name,
                                    '--location', location,
-                                   '--source', target_blob_path],
+                                   '--source', target_blob_path,
+                                   '--source-storage-account-id', source_storage_account_id],
                                   subscription=target_subscription)
 
     json_output = run_cli_command(cli_cmd, return_as_json=True)
@@ -125,9 +138,9 @@ def create_target_image(location, transient_resource_group_name, source_type, so
 
     # Optionally create the final image
     if export_as_snapshot:
-        logger.warn("%s - Skipping image creation", location)
+        logger.warning("%s - Skipping image creation", location)
     else:
-        logger.warn("%s - Creating final image", location)
+        logger.warning("%s - Creating final image", location)
         if target_name is None:
             target_image_name = source_object_name
             if source_type != 'image':
@@ -170,7 +183,7 @@ def wait_for_blob_copy_operation(blob_name, target_container_name, target_storag
         if current_progress != prev_progress:
             msg = "{0} - Copy progress: {1}%"\
                 .format(location, str(current_progress))
-            logger.warn(msg)
+            logger.warning(msg)
 
         prev_progress = current_progress