+++ src/vm-repair/azext_vm_repair/tests/latest/test_repair_commands.py
@@ -3,10 +3,10 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
+from azure.cli.testsdk import LiveScenarioTest, ResourceGroupPreparer
 
 
-class WindowsManagedDiskCreateRestoreTest(ScenarioTest):
+class WindowsManagedDiskCreateRestoreTest(LiveScenarioTest):
 
     @ResourceGroupPreparer(location='westus2')
     def test_vmrepair_WinManagedCreateRestore(self, resource_group):
@@ -16,15 +16,15 @@ class WindowsManagedDiskCreateRestoreTest(ScenarioTest):
 
         # Create test VM
         self.cmd('vm create -g {rg} -n {vm} --admin-username azureadmin --image Win2016Datacenter --admin-password !Passw0rd2018')
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         # Something wrong with vm create command if it fails here
         assert len(vms) == 1
 
         # Test create
-        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018').get_output_in_json()
+        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018 -o json').get_output_in_json()
 
         # Check repair VM
-        repair_vms = self.cmd('vm list -g {}'.format(result['repair_resouce_group'])).get_output_in_json()
+        repair_vms = self.cmd('vm list -g {} -o json'.format(result['repair_resource_group'])).get_output_in_json()
         assert len(repair_vms) == 1
         repair_vm = repair_vms[0]
         # Check attached data disk
@@ -34,12 +34,12 @@ class WindowsManagedDiskCreateRestoreTest(ScenarioTest):
         self.cmd('vm repair restore -g {rg} -n {vm} --yes')
 
         # Check swapped OS disk
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         source_vm = vms[0]
         assert source_vm['storageProfile']['osDisk']['name'] == result['copied_disk_name']
 
 
-class WindowsUnmanagedDiskCreateRestoreTest(ScenarioTest):
+class WindowsUnmanagedDiskCreateRestoreTest(LiveScenarioTest):
 
     @ResourceGroupPreparer(location='westus2')
     def test_vmrepair_WinUnmanagedCreateRestore(self, resource_group):
@@ -49,15 +49,15 @@ class WindowsUnmanagedDiskCreateRestoreTest(ScenarioTest):
 
         # Create test VM
         self.cmd('vm create -g {rg} -n {vm} --admin-username azureadmin --image Win2016Datacenter --admin-password !Passw0rd2018 --use-unmanaged-disk')
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         # Something wrong with vm create command if it fails here
         assert len(vms) == 1
 
         # Test create
-        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018').get_output_in_json()
+        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018 -o json').get_output_in_json()
 
         # Check repair VM
-        repair_vms = self.cmd('vm list -g {}'.format(result['repair_resouce_group'])).get_output_in_json()
+        repair_vms = self.cmd('vm list -g {} -o json'.format(result['repair_resource_group'])).get_output_in_json()
         assert len(repair_vms) == 1
         repair_vm = repair_vms[0]
         # Check attached data disk
@@ -67,12 +67,12 @@ class WindowsUnmanagedDiskCreateRestoreTest(ScenarioTest):
         self.cmd('vm repair restore -g {rg} -n {vm} --yes')
 
         # Check swapped OS disk
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         source_vm = vms[0]
         assert source_vm['storageProfile']['osDisk']['vhd']['uri'] == result['copied_disk_uri']
 
 
-class LinuxManagedDiskCreateRestoreTest(ScenarioTest):
+class LinuxManagedDiskCreateRestoreTest(LiveScenarioTest):
 
     @ResourceGroupPreparer(location='westus2')
     def test_vmrepair_LinuxManagedCreateRestore(self, resource_group):
@@ -82,15 +82,15 @@ class LinuxManagedDiskCreateRestoreTest(ScenarioTest):
 
         # Create test VM
         self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --admin-username azureadmin --admin-password !Passw0rd2018')
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         # Something wrong with vm create command if it fails here
         assert len(vms) == 1
 
         # Test create
-        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018').get_output_in_json()
+        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018 -o json').get_output_in_json()
 
         # Check repair VM
-        repair_vms = self.cmd('vm list -g {}'.format(result['repair_resouce_group'])).get_output_in_json()
+        repair_vms = self.cmd('vm list -g {} -o json'.format(result['repair_resource_group'])).get_output_in_json()
         assert len(repair_vms) == 1
         repair_vm = repair_vms[0]
         # Check attached data disk
@@ -100,12 +100,12 @@ class LinuxManagedDiskCreateRestoreTest(ScenarioTest):
         self.cmd('vm repair restore -g {rg} -n {vm} --yes')
 
         # Check swapped OS disk
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         source_vm = vms[0]
         assert source_vm['storageProfile']['osDisk']['name'] == result['copied_disk_name']
 
 
-class LinuxUnmanagedDiskCreateRestoreTest(ScenarioTest):
+class LinuxUnmanagedDiskCreateRestoreTest(LiveScenarioTest):
 
     @ResourceGroupPreparer(location='westus2')
     def test_vmrepair_LinuxUnmanagedCreateRestore(self, resource_group):
@@ -115,15 +115,15 @@ class LinuxUnmanagedDiskCreateRestoreTest(ScenarioTest):
 
         # Create test VM
         self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --admin-username azureadmin --admin-password !Passw0rd2018 --use-unmanaged-disk')
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         # Something wrong with vm create command if it fails here
         assert len(vms) == 1
 
         # Test create
-        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018').get_output_in_json()
+        result = self.cmd('vm repair create -g {rg} -n {vm} --repair-username azureadmin --repair-password !Passw0rd2018 -o json').get_output_in_json()
 
         # Check repair VM
-        repair_vms = self.cmd('vm list -g {}'.format(result['repair_resouce_group'])).get_output_in_json()
+        repair_vms = self.cmd('vm list -g {} -o json'.format(result['repair_resource_group'])).get_output_in_json()
         assert len(repair_vms) == 1
         repair_vm = repair_vms[0]
         # Check attached data disk
@@ -133,12 +133,12 @@ class LinuxUnmanagedDiskCreateRestoreTest(ScenarioTest):
         self.cmd('vm repair restore -g {rg} -n {vm} --yes')
 
         # Check swapped OS disk
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         source_vm = vms[0]
         assert source_vm['storageProfile']['osDisk']['vhd']['uri'] == result['copied_disk_uri']
 
 
-class WindowsRunHelloWorldTest(ScenarioTest):
+class WindowsRunHelloWorldTest(LiveScenarioTest):
 
     @ResourceGroupPreparer(location='westus2')
     def test_vmrepair_WinRunHelloWorld(self, resource_group):
@@ -148,18 +148,18 @@ class WindowsRunHelloWorldTest(ScenarioTest):
 
         # Create test VM
         self.cmd('vm create -g {rg} -n {vm} --admin-username azureadmin --image Win2016Datacenter --admin-password !Passw0rd2018')
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         # Something wrong with vm create command if it fails here
         assert len(vms) == 1
 
         # Test create
-        result = self.cmd('vm repair run -g {rg} -n {vm} --run-id win-hello-world').get_output_in_json()
+        result = self.cmd('vm repair run -g {rg} -n {vm} --run-id win-hello-world -o json').get_output_in_json()
 
         # Check Output
         assert 'Hello World!' in result['output']
 
 
-class LinuxRunHelloWorldTest(ScenarioTest):
+class LinuxRunHelloWorldTest(LiveScenarioTest):
 
     @ResourceGroupPreparer(location='westus2')
     def test_vmrepair_LinuxRunHelloWorld(self, resource_group):
@@ -169,12 +169,12 @@ class LinuxRunHelloWorldTest(ScenarioTest):
 
         # Create test VM
         self.cmd('vm create -g {rg} -n {vm} --admin-username azureadmin --image UbuntuLTS --admin-password !Passw0rd2018')
-        vms = self.cmd('vm list -g {rg}').get_output_in_json()
+        vms = self.cmd('vm list -g {rg} -o json').get_output_in_json()
         # Something wrong with vm create command if it fails here
         assert len(vms) == 1
 
         # Test create
-        result = self.cmd('vm repair run -g {rg} -n {vm} --run-id linux-hello-world').get_output_in_json()
+        result = self.cmd('vm repair run -g {rg} -n {vm} --run-id linux-hello-world -o json').get_output_in_json()
 
         # Check Output
         assert 'Hello World!' in result['output']